requirements.txt
chris_plugin==0.4.0
python-chrisclient==2.11.1
loguru